installer: Remove todo comments about downloading latest release
The installer embeds the release assets corresponding to its own
version. This removes downloads that have to happen at runtime (the
assets are downloaded as part of the installer). Plus, it might not make
sense to install newer assets using an older installer, since the way
installation works may change.